refactor: add in_memory context
