+++ tap_integration_tests/Cargo.toml
@@ -9,7 +9,7 @@ publish = false
 
 [dependencies]
 tap_aggregator = { version = "0.2.0", path = "../tap_aggregator" }
-tap_core = { version = "0.7.0", path = "../tap_core" }
+tap_core = { version = "0.7.0", path = "../tap_core", features = ["in_memory"]}
 jsonrpsee = { version = "0.18.0", features = ["http-client", "server"] }
 ethers = "2.0.0"
 clap = { version = "4.2.4", features = ["derive", "env"] }
