+++ tap_integration_tests/tests/indexer_mock.rs
