+++ tap_integration_tests/tests/showcase.rs
@@ -22,20 +22,19 @@ use jsonrpsee::{
 use rand::{rngs::StdRng, Rng, SeedableRng};
 use rstest::*;
 
-use common::{ExecutorMock, QueryAppraisals};
 use tap_aggregator::{jsonrpsee_helpers, server as agg_server};
 use tap_core::{
+    manager::context::memory::{checks::get_full_list_of_checks, *},
     rav::SignedRAV,
     receipt::{
-        checks::{mock::get_full_list_of_checks, Checks, TimestampCheck},
+        checks::{Checks, TimestampCheck},
         Receipt,
     },
     signed_message::{EIP712SignedMessage, MessageId},
     tap_eip712_domain,
 };
 
-mod common;
-mod indexer_mock;
+use crate::indexer_mock;
 
 // Fixtures for sender aggregator server
 #[fixture]
